seventh commit

+++ c/function.h
@@ -1,4 +1,4 @@
 #include <stdio.h>
 
 void calcul(int a, int b);
-int factoriel(int n);
+long factoriel(int n);
